requirements.txt
tushare>=1.2.89
pandas>=1.3.0
numpy>=1.21.0
openai>=1.0.0
akshare>=1.10.0
ta>=0.10.0
streamlit>=1.0.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
requests>=2.31.0
schedule>=1.2.0 